SocketIO Flask server with celery log processing and reporting on upload
The celery task reports status via websockets to inform the user of the
processing state. When complete bokeh can be used to perform analysis on
the log. 

Due to memory errors in the standard bokeh server I have had to wrap
them and spin them up as requested by the user. This helps keep memory
usage on the webserver down. Currently there is no detection method to
know if the bokeh server is ready to accept connections. I will need to
harden this along with adding password and config support for redis.